requirements.txt
molo.core<7.0.0,>=6.0.0
celery<4.0
django-celery
wagtailsurveys==0.1.1
wagtail-personalisation-molo==0.10.3
django==1.10.8
psycopg2
html5lib==0.9999999
Unidecode==0.04.16
